Move `WebpackPrecompiledLoader`
